requirements-ja.txt
pyspark==3.2.1
fire==0.3.1
pyyaml
bs4
html2text
python-stdnum
numpy
SudachiPy==0.5.4
SudachiDict-core